migrate oidcConformant flag into Auth0 class
